Moving haltIstioSidecar() deferrec call to main Run() function
